==> simulation/ptanh_simulation/1_act_output.ipynb <==
import pandas
import numpy as np
import matplotlib.pyplot as plt
from scipy import optimize as op
import math
import torch

data = pandas.read_csv('../result/Vout_10000_ptanh_zero_cross_pNN_ACT_interactive1.csv')
data

# 无用列

f = data.iloc[:,1:]
f

a = torch.tensor(f.to_numpy()).T
a.shape

a

np.savetxt('./data/act_output.csv', a, fmt='%f', delimiter=',')
