==> simulation/ptanh_simulation/2_act_RT.ipynb <==
import pandas
import numpy as np
import matplotlib.pyplot as plt
from scipy import optimize as op
import math
import torch

data = pandas.read_csv('../result/Vout_10000_ptanh_zero_cross_pNN_ACT_interactive1.csv',nrows= 0)
data

datalist=data.columns.values.tolist()
print(datalist)
print(len(datalist))

result = []
for j in range(1,10001):
    a = datalist[j].split(' ')
    nums = []
    for k in range(2, 13, 2):
        nums.append(a[k])
    b=map(float,nums)    
    result.append(list(b))

result_temp = torch.tensor(result)
result_temp.max(0)[0]

result_temp.min(0)[0]

c = torch.tensor(result)
c.shape

np.savetxt('./data/act_RT.csv', c, fmt='%f', delimiter=',')

r1 = c[:, 0].view(-1,1)
r2 = c[:, 1].view(-1,1)
w1 = c[:, 2].view(-1,1)
l1 = c[:, 3].view(-1,1)
w2 = c[:, 4].view(-1,1)
l2 = c[:, 5].view(-1,1)
ratio_1 = r2/r1
ratio_2 = l1/w1
ratio_3 = l2/w2
r1.shape, r2.shape, w1.shape, l1.shape, w2.shape, l2.shape

C = torch.cat([r1, r2, w1, l1, w2, l2, ratio_1, ratio_2, ratio_3], dim=1)
C.shape

np.savetxt('./data/act_RT_ratio.csv', C, fmt='%f', delimiter=',')
