==> simulation/ptanh_simulation/3_act_tanh.ipynb <==
%pylab inline 
import matplotlib.pyplot as plt 
import numpy as np

import pandas
import numpy as np
import matplotlib.pyplot as plt
from scipy import optimize as op
import math
import torch

def f(x, eta1,eta2,eta3,eta4):
    return eta1 + eta2 * np.tanh((x - eta3) * eta4)

data1 = pandas.read_csv("./data/act_output.csv", header=None)

data1.shape

data1

result = []
x=np.linspace(-2,2,4001)
for k in range(10000):
    print(k)
    plt.figure()
    f1 = data1.iloc[k, :]
    y = np.array([float(f2) for f2 in f1.to_numpy()]) 
    eta1,eta2,eta3,eta4= op.curve_fit(f, x, y, maxfev=1000000,p0=[0.,1.,0.,1.], bounds=([-np.inf, 0, -np.inf, 0],[np.inf, np.inf, np.inf, np.inf]))[0]
    result.append([eta1, eta2, eta3, eta4])
    plt.plot(x, y,color='green')
    ys = eta1 + eta2 * np.tanh((x - eta3) * eta4)
    plt.plot(x, ys,color='red')
    
    plt.savefig(f'./act_eta_figures/{k}.pdf', format='pdf', bbox_inches='tight')
    plt.show()

len(result)

a = torch.tensor(result)
a.shape

np.savetxt('./data/act_eta.csv', a, fmt='%f', delimiter=',')
