==> simulation/ptanh_simulation/4.2_act_dataset_ratio.ipynb <==
import os
import math
import torch
import random
import pickle
import pandas
import calendar
import numpy as np
import matplotlib.pyplot as plt
from torch.utils.data import TensorDataset, DataLoader
os.environ['KMP_DUPLICATE_LIB_OK']='True'
torch.set_default_tensor_type(torch.FloatTensor)
import config

# Read and process data

x = pandas.read_csv("./data/act_RT_ratio.csv", header=None)

X = []
for i in range(10000):
    
    f1 = x.iloc[i, :]
    F = [float(f) for f in f1.to_numpy()]
    X.append(F)
X = torch.tensor(X)
X.shape

X_max = torch.max(X, dim=0)[0]
X_min = torch.min(X, dim=0)[0]

X_max

X_min

Xn = config.Normalization(X, X_min, X_max)
Xn

Xn.max(), Xn.min()

y = pandas.read_csv("./data/act_eta.csv", header=None)
y.shape

Y = []
for i in range(10000):
    
    f1 = y.iloc[i, :]
    F = [float(f) for f in f1.to_numpy()]
    Y.append(F)
Y = torch.tensor(Y)
Y.shape

Y_max = torch.max(Y, dim=0)[0]
Y_min = torch.min(Y, dim=0)[0]

Y_min

Y_max

Yn = config.Normalization(Y, Y_min, Y_max)

Yn.max(), Yn.min()

# Make dataset

E_train = int(X.shape[0]*0.7)
E_valid = int(X.shape[0]*0.2)
E_test  = int(X.shape[0] - E_train - E_valid)

config.SetSeed()
index = torch.randperm(X.shape[0])
index_learn = index[:E_train+E_valid]
index_train = index[:E_train]
index_valid = index[E_train:E_train+E_valid]
index_test  = index[-E_test:]

X_learn, Y_learn = X[index_learn,:], Y[index_learn,:]
X_train, Y_train = X[index_train,:], Y[index_train,:]
X_valid, Y_valid = X[index_valid,:], Y[index_valid,:]
X_test,  Y_test  = X[index_test,:] , Y[index_test,:]

Xn_learn, Yn_learn = Xn[index_learn,:], Yn[index_learn,:]
Xn_train, Yn_train = Xn[index_train,:], Yn[index_train,:]
Xn_valid, Yn_valid = Xn[index_valid,:], Yn[index_valid,:]
Xn_test,  Yn_test  = Xn[index_test,:] , Yn[index_test,:]

a = {'Xn': Xn, 'Yn': Yn, 
     'X': X,   'Y': Y, 
     
     'Xn_learn': Xn_learn, 'Yn_learn': Yn_learn, 
     'Xn_train': Xn_train, 'Yn_train': Yn_train, 
     'Xn_valid': Xn_valid, 'Yn_valid': Yn_valid, 
     'Xn_test' : Xn_test,  'Yn_test' : Yn_test,
     
     'X_learn': X_learn, 'Y_learn': Y_learn, 
     'X_train': X_train, 'Y_train': Y_train, 
     'X_valid': X_valid, 'Y_valid': Y_valid, 
     'X_test' : X_test,  'Y_test' : Y_test,
     
     'X_max':X_max, 'X_min':X_min,
     'Y_max':Y_max, 'Y_min':Y_min,}

with open('./data/act_dataset.p', 'wb') as file:
    pickle.dump(a, file)
with open('../../LNC/dataset/act_dataset.p', 'wb') as file:
    pickle.dump(a, file)

a = Y[:,0]
b = Y[:,1]
c = Y[:,2]
d = Y[:,3]

re = c*d

torch.where(re == re.min())

xxx = Xn[2181,:6]

torch.log(xxx / (1-xxx))
